small stuff before the presentation

+++ backend/api/models.py
@@ -86,18 +86,22 @@ class TileReactionComponent(models.Model):
         db_table = "tiles"
 
 class TileSubsystem(models.Model):
-    subsystem_id = models.FloatField()
+    subsystem_id = models.IntegerField()
     subsystem_name = models.CharField(max_length=200, null=False)
     compartment_name = models.CharField(max_length=50, null=False)
-    x_top_left = models.FloatField(null=False)
-    y_top_left = models.FloatField(null=False)
-    x_bottom_right = models.FloatField(null=False)
-    y_bottom_right = models.FloatField(null=False)
+    x_top_left = models.IntegerField(null=False)
+    y_top_left = models.IntegerField(null=False)
+    x_bottom_right = models.IntegerField(null=False)
+    y_bottom_right = models.IntegerField(null=False)
 
     class Meta:
         db_table = "tile_subsystems"
-#create table tile_subsystems(subsystem_id integer not null, subsystem_name varchar(200), compartment_name varchar(50), x_top_left integer, y_top_left integer, x_bottom_right integer, y_bottom_right integer)
-#insert into tile_subsystems values(38, 'Tricarboxylic acid cycle and glyoxylate/dicarboxylate metabolism', 'm', 12000, 7000, 13000, 8000);
+#create table tile_subsystems(id integer, subsystem_id integer not null, subsystem_name varchar(200), compartment_name varchar(50), x_top_left integer, y_top_left integer, x_bottom_right integer, y_bottom_right integer)
+#insert into tile_subsystems values(1, 38, 'Tricarboxylic acid cycle and glyoxylate/dicarboxylate metabolism', 'm', 12000, 7000, 13000, 8000);
+#insert into tile_subsystems values(2, 38, 'Tricarboxylic acid cycle and glyoxylate/dicarboxylate metabolism', 'm', 9000, 7000, 16000, 12000);
+#insert into tile_subsystems values(3, 38, 'Tricarboxylic acid cycle and glyoxylate/dicarboxylate metabolism', 'm', 10000, 5000, 17000, 10000);
+#insert into tile_subsystems values(4, 38, 'Tricarboxylic acid cycle and glyoxylate/dicarboxylate metabolism', 'm', 10500, 4700, 16500, 10500);
+#insert into tile_subsystems values(4, 38, 'Tricarboxylic acid cycle and glyoxylate/dicarboxylate metabolism', 'm', 12800, 4600, 16800, 9000);
 
 ##########################################################################################################################
 ##########################################################################################################################
